chore: bump version to 0.16.8 and tidy tests/clippy
- remove cfg(test) App::new constructors from construct.rs
- fix clippy field_reassign_with_default / too_many_arguments / let-else lints
- stabilize announced-shutdown assertion to wait on the event
- move audiobookshelf render tests to end of file

+++ crates/mbv-core/src/playback_queue_tests.rs
@@ -44,7 +44,7 @@ fn audiobookshelf_episode(library_item_id: &str, episode_id: &str) -> Audiobooks
         show_title: Some("Show".into()),
         author: Some("Author".into()),
         duration_ticks: Some(120 * TICKS_PER_SECOND as u64),
-        position_ticks: 30 * TICKS_PER_SECOND as i64,
+        position_ticks: 30 * TICKS_PER_SECOND,
         played: false,
         pub_date_secs: Some(1_700_000_000),
         is_finished: false,
